change file and create gameboard controller js
